moved all code for csv parsing, code and sqlite gen plus db loading has been moved into the csv_converter crate
