+++ csv_converter/Cargo.toml
@@ -12,4 +12,5 @@ regex = "1.0.0"
 serde = "1.0.70"
 serde_derive = "1.0.66"
 serde_json = "1.0.19"
+sqlite = "0.23.9"
 toml = "0.4"
